test succeed
after deleted /version/1.6.4, Minecraft work perfectly.
checkout files and end the test#1

PS1 after running, found a folder named ".Minecraft". it will update
everytime when loading mods. Maybe it can't find /.Minecraft so created
itself. if the modpack run on other computers, this problem probably will
disappear. so add it to .gitignore .

PS2 there is a puzzling file(or dir) named \357\200\233
don't know what it is, so add it to repo