Install IndustrialCraft2
install ic2 and its API successfully. no isssues have found so far.

there is three files, "mod file","deve","api"on the download page of IC2.
I can't tell which is nessensarry, so I installed "modfile" and API.
maybe i should open a test to test which to install is the best.